requirements.txt
click==8.1.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
scipy==1.11.4
six==1.16.0
